feat: Add NLP stop words and TokenService
Introduces a new configuration for NLP stop words in `AGI.php` to filter common words and HTML tags during keyword extraction. This aims to improve the quality of identified keywords by the `TokenService`.

Also, registers the `TokenService` in `Services.php` to make it available application-wide. A placeholder function `_classifyIntent` is added to `GeminiController.php` to demonstrate where intent classification logic would be integrated in the future.