feat: Add new service definitions to Config/Services.php
This commit introduces the definitions for several new service classes within the `app/Config/Services.php` file.

The following services have been added:
- PaystackService
- CryptoService
- GeminiService
- RecaptchaService

This change allows these services to be easily instantiated and managed as singletons or new instances throughout the application, following the existing pattern for service configuration. This is a necessary step for integrating these new functionalities into the application.