feat: Disable font subsetting in DomPDF for WASI compatibility
This change disables font subsetting in DomPDF by setting 'isFontSubsettingEnabled' to false. This is necessary to improve compatibility with the restricted WASI file system, as it prevents DomPDF from attempting to parse and extract individual characters from font files. Instead, the entire font file will be embedded, which is a simpler and more reliable approach in this environment.